move performance_metrics
