+++ aeon/benchmarking/performance_metrics/anomaly_detection/_vus_metrics.py
@@ -15,7 +15,7 @@ import warnings
 
 import numpy as np
 
-from aeon.performance_metrics.anomaly_detection._util import check_y
+from aeon.benchmarking.performance_metrics import check_y
 
 
 def _anomaly_bounds(y_true: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
