+++ aeon/benchmarking/performance_metrics/segmentation/__init__.py
@@ -1,6 +1,6 @@
 """Metrics to assess performance of segmentation task."""
 
-from aeon.performance_metrics.segmentation.metrics import (  # noqa
+from aeon.benchmarking.performance_metrics import (  # noqa
     count_error,
     hausdorff_error,
     prediction_ratio,
