+++ aeon/benchmarking/performance_metrics/tests/test_stats.py
@@ -7,9 +7,13 @@ from numpy.testing import assert_almost_equal
 from scipy.stats import rankdata
 
 import aeon
+from aeon.benchmarking.performance_metrics import (
+    check_friedman,
+    nemenyi_test,
+    wilcoxon_test,
+)
 from aeon.benchmarking.results_loaders import get_estimator_results_as_array
 from aeon.datasets.tsc_datasets import univariate_equal_length
-from aeon.performance_metrics.stats import check_friedman, nemenyi_test, wilcoxon_test
 
 data_path = os.path.join(
     os.path.dirname(aeon.__file__),
